+++ lib/Models/SensorObservationServiceCatalogItem.ts
@@ -26,12 +26,11 @@ import TableStyleTraits from "../Traits/TableStyleTraits";
 import CreateModel from "./CreateModel";
 import createStratumInstance from "./createStratumInstance";
 import proxyCatalogItemUrl from "./proxyCatalogItemUrl";
-import { AvailableStyle } from "./SelectableStyle";
 import StratumFromTraits from "./StratumFromTraits";
 import StratumOrder from "./StratumOrder";
 import Terria from "./Terria";
-import TableColumnTraits from "../Traits/TableColumnTraits";
 import CommonStrata from "./CommonStrata";
+import { SelectableDimension } from "./SelectableDimensions";
 import { BaseModel } from "./Model";
 
 interface GetFeatureOfInterestResponse {
@@ -547,12 +546,7 @@ export default class SensorObservationServiceCatalogItem extends TableMixin(
   }
 
   @computed
-  get styleSelector() {
-    return undefined;
-  }
-
-  @computed
-  get styleSelectors() {
+  get selectableDimensions() {
     return filterOutUndefined([
       this.proceduresSelector,
       this.observablesSelector
@@ -560,8 +554,8 @@ export default class SensorObservationServiceCatalogItem extends TableMixin(
   }
 
   @computed
-  get proceduresSelector() {
-    const proceduresSelector = super.styleSelector;
+  get proceduresSelector(): SelectableDimension | undefined {
+    const proceduresSelector = super.styleDimensions;
     if (proceduresSelector === undefined) return;
 
     const item = this;
@@ -574,7 +568,7 @@ export default class SensorObservationServiceCatalogItem extends TableMixin(
   }
 
   @computed
-  get observablesSelector() {
+  get observablesSelector(): SelectableDimension | undefined {
     if (this.mapItems.length === 0) {
       return;
     }
@@ -586,7 +580,7 @@ export default class SensorObservationServiceCatalogItem extends TableMixin(
       get name(): string {
         return item.observablePropertiesName;
       },
-      get availableStyles(): readonly AvailableStyle[] {
+      get options() {
         return filterOutUndefined(
           item.observableProperties.map(p => {
             if (p.identifier && p.title) {
@@ -598,10 +592,10 @@ export default class SensorObservationServiceCatalogItem extends TableMixin(
           })
         );
       },
-      get activeStyleId(): string | undefined {
+      get selectedId(): string | undefined {
         return item.selectedObservableId;
       },
-      chooseActiveStyle(stratumId: string, observableId: string) {
+      setDimensionValue(stratumId: string, observableId: string) {
         item.setTrait(stratumId, "selectedObservableId", observableId);
       }
     };